Changed task_plan publish API to the new format (no background jobs yet)

+++ app/routines/distribute_tasks.rb
@@ -1,34 +1,18 @@
-require 'json-schema'
-
 class DistributeTasks
 
-  lev_routine
+  lev_routine transaction: :no_transaction # The controller already wraps this in a transaction
 
   uses_routine IndividualizeTaskingPlans, as: :get_tasking_plans
 
   protected
 
-  def validate_json(schema, object, options = {})
-    options[:insert_defaults] = true if options[:insert_defaults].nil?
-
-    JSON::Validator.fully_validate(schema, object, options)
-  end
-
   def exec(task_plan)
-    owner = task_plan.owner
+    task_plan.publish_last_requested_at = Time.now
+
     assistant = task_plan.assistant
 
-    # Validate the given settings against the assistant's schema
-    # Intervention settings already included when the task_plan was saved
-    err = validate_json(assistant.schema, task_plan.settings)
-
-    fatal_error(code: :invalid_settings, message: 'Invalid settings', data: err) unless err.empty?
-
     # Delete pre-existing assignments
-    unless task_plan.tasks.empty?
-      task_plan.tasks.destroy_all
-      task_plan.reload
-    end
+    task_plan.tasks.destroy_all unless task_plan.tasks.empty?
 
     tasking_plans = run(:get_tasking_plans, task_plan).outputs.tasking_plans
 
@@ -53,7 +37,8 @@ class DistributeTasks
 
     outputs[:tasks] = tasks
 
-    task_plan.update_attributes(published_at: Time.now)
+    task_plan.published_at = Time.now
+    task_plan.save
   end
 
 end
